requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.0.2
